### 2019-02-23 #1275 @arfon
:wave: @jgostick - thanks for your submission to JOSS. There are a few issues that need fixing before we can proceed to assign a handling editor:

1. Please add an open source license to this repository
2. Please write a paper to accompany your submission. [This section](https://joss.readthedocs.io/en/latest/submitting.html#what-should-my-paper-contain) of our docs should help.
