### 2019-02-25 #1275 @jgostick
I'm sorry, I thought I'd done this.  I added all this to the Master branch, but the default branch on github was the dev branch, so perhaps that was the issues?  Anyway, I have changed the default to master and updated the version number to V1.0.1.  Do I need to tell the system to look at this new tag?  If so, how?
