### 2019-02-25 #1275 @danielskatz
There is some extra space before the name associated with affil 2.  Please remove it.  You can also regenerate the pdf yourself by entering ```@whedon generate pdf``` as a new comment here.
