### 2019-02-25 e2b8e4e
Fixing white space error in paper.md
