### 2019-02-25 #1275 @danielskatz
The third-to-last reference could use some work on the author names.
