### 2019-02-25 5d5cfcd
Final tweaks to paper info
