+++ paper.bib
@@ -50,7 +50,7 @@
   journaltitle = {Journal of Microscopy},
   date = {2016-09},
   pages = {280-292},
-  author = {TAIWO, OLUWADAMILOLA O. and FINEGAN, DONAL P. and EASTWOOD, DAVID S. and FIFE, JULIE L. and BROWN, LEON D. and DARR, JAWWAD A. and LEE, PETER D. and BRETT, DANIEL J.L. and SHEARING, PAUL R.},
+  author = {Taiwo, O.O. and Finegan, D.P. and Eastwood, D.S. and Fife, J.L. and Brown, L.D. and Darr, J.A. and Lee, P.D. and Brett, D.J.L. and Shearing, PR.},
 }
 
 @article{garcia-salaberriAnalysisRepresentativeElementary2018,
