+++ paper.md
@@ -46,9 +46,9 @@ bibliography: paper.bib
 
 # Summary
 
-Porous materials play a central role in many technologies, from micron-thick electrodes used in batteries to kilometer-long geologic formations of interest in oil recovery and contaminant transport in aquifers.  These applications share a common interest in analyzing the transport processes occurring at the pore-scale, since these ultimately control the observable macroscopic behavior.  Images of porous materials are a valuable source of information, since performance and structure are intimately linked through the topology and geometry of the media.  A variety of techniques are available for imaging a material's internal pore structure with sub-micron resolution, including X-ray tomography [@colesPoreLevelImaging1998], ptychography [@dierolfPtychographicXrayComputed2010], and FIB-SEM sectioning [@heaneyFocusedIonBeam2001].  Each of these tools can provide exquisitely detailed information in the form of 3D volumetric images, and retrieving quantitative information from these images has become a vital tool in all areas of porous media research.
+Porous materials play a central role in many technologies, from micron-thick electrodes used in batteries to kilometer-long geologic formations of interest in oil recovery and contaminant transport in aquifers.  These applications share a common interest in analyzing the transport processes occurring at the pore-scale, since these ultimately control the observable macroscopic behavior.  Images of porous materials are a valuable source of information, since performance and structure are intimately linked through the topology and geometry of the media.  A variety of techniques are available for imaging a material's internal pore structure with sub-micron resolution, including X-ray tomography [@colesPoreLevelImaging1998], ptychography [@dierolfPtychographicXrayComputed2010], and FIB-SEM sectioning [@heaneyFocusedIonBeam2001].  Each of these tools can provide exquisitely detailed images, and retrieving quantitative information from these images has become a vital tool in all areas of porous media research.
 
-The analysis of images, both 2D sections and 3D volumetric data [@taiwoComparisonThreeDimensional2016], is an established field, dating to the advent of computers with sufficient memory to store and manipulate the rather large file sizes.  Pioneering work by Torquato, summarized in his textbook [@torquatoRandomHeterogeneousMaterials2005], laid a solid foundation by defining many metrics still widely used, including chord-length distribution [@jervisXrayNanoComputed2018] and radial-density function [@kokInsightsEffectStructural2018].  The field is far from dormant, however, as increasingly powerful computers allow application of concomitantly advanced algorithms, such as watershed partitioning to create networks [@khanDualNetworkExtraction2019, @gostickVersatileEfficientPore2017] for pore-network modeling [@gostickOpenPNMPoreNetwork2016], and direct numerical simulations [@garcia-salaberriImplicationsInherentInhomogeneities2019] which can often supplant experiments [@garcia-salaberriAnalysisRepresentativeElementary2018] or even conduct experiments that are not physically possible [@garcia-salaberriEffectiveDiffusivityPartiallysaturated2015, @kokMassTransferFibrous2019].  The aim of PoreSpy is to collect the various tools and functions needed by researchers to perform quantitative analysis of volumetric images of porous materials into a single, easy to use, open-source package, written in the popular Python language which contains several image processing libraries offering the full suite of necessary tools [@gouillartAnalyzingMicrotomographyData2016, @scikit-image].  PoreSpy contains a large, though not exhaustive, set of the most commonly used tools for porous media analysis, which typically work by applying a series of the basic image analysis functions. The functions in PoreSpy are organized into the following categories:
[…]
