### 2019-03-04 #1296 @whedon
[REVIEW]: PoreSpy: A Python Toolkit for Quantitative Analysis of Porous Media Images
**Submitting author:** @jgostick (<a href="http://orcid.org/0000-0001-7736-7124">Jeff Gostick</a>)
**Repository:** <a href="https://github.com/PMEAL/porespy" target ="_blank">https://github.com/PMEAL/porespy</a>
**Version:** V1.1.2
