### 2019-03-08 28571d3
updating conda build on travis
