+++ .travis.yml
@@ -10,6 +10,7 @@ before_install:
   - sh -e /etc/init.d/xvfb start
   - wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh
   - bash miniconda.sh -b -p $HOME/miniconda
+  - conda config --add channels conda-forge
   - export PATH="$HOME/miniconda/bin:$PATH"
 
 install:
