+++ .travis.yml
@@ -10,8 +10,8 @@ before_install:
   - sh -e /etc/init.d/xvfb start
   - wget https://repo.continuum.io/miniconda/Miniconda3-latest-Linux-x86_64.sh -O miniconda.sh
   - bash miniconda.sh -b -p $HOME/miniconda
-  - conda config --add channels conda-forge
   - export PATH="$HOME/miniconda/bin:$PATH"
+  - conda config --add channels conda-forge
 
 install:
   - conda install --yes python=$TRAVIS_PYTHON_VERSION --file requirements/conda_requirements.txt
