+++ porespy/network_extraction/__utils__.py
@@ -1,0 +1,2 @@
+class _net_dict(dict):  # assign out values to dummy dict
+    pass
