### 2019-03-08 25553c8
adding net_dict to utils
