+++ porespy/networks/__init__.py
@@ -30,5 +30,6 @@ from .__funcs__ import add_boundary_regions
 from .__funcs__ import map_to_regions
 from .__funcs__ import generate_voxel_image
 from .__getnet__ import regions_to_network
+from .__utils__ import _net_dict
 from .__snow__ import snow
 from .__snow_dual__ import snow_dual
