### 2019-03-08 #1296 @jgostick
Not sure if it's ok to interject in the review process, but this seems relevant to the review process:

Sometime between submitting to JOSS and now, Numpy (1.16) made a change that breaks Skimage (0.14.1)  The Skimage team has already fixed this in 0.14.2, but this version is NOT currently the default on the Conda distribution.  So, if a user were to install Conda, they get a version of Numpy and Skimage that do not work together.  The latest version of Skimage is available on the conda-forge channel, but not sure how long it will be until it reaches the default Conda channel.

In our package we try not to specify hard version numbers, to avoid users needing to make virtualenvs for each package they want to use, so at this time the install may not work, depending the version of conda is being used.
