### 2019-03-09 8551ae0
adding readme to examples folder, updating main readme url to examples
