+++ porespy/__init__.py
@@ -79,7 +79,7 @@ Use some filters from PoreSpy:
 
 '''
 
-__version__ = "1.0.2"
+__version__ = "1.0.1"
 
 from . import metrics
 from . import tools
