### 2019-03-12 #1296 @yxqd
Hi @jgostick. I was able to install porespy with conda and will test its functionality tomorrow. Here are some questions regarding the paper itself:

* Should [Matthew Lam](https://github.com/mattrlam) be a co-author? Looking at commits only, he seems to contribute similarly to Matthew Kok and Mohammadamin Sadeghi.
* Could you please add more references to the first sentence of the paper? For example, https://doi.org/10.2138/rmg.2015.80.04 may be a good one
* There is one reference for your paper missing DOI
