### 2019-03-12 #1296 @jgostick
HI @yxqd
I'm glad you were able to get it installed...but if you haven't tried importing it you might run into the above mentioned problem.  You can upgrade skimage with ``conda install scikit-image -c conda-forge``

Regarding your questions:
* Matt Lam was an undergrad student that worked on a random walk algorithm.  We since removed this and put it into a separate package called [PyTrax](https://github.com/PMEAL/pytrax).  We submitted a paper to SoftwareX with his name on it, so he's getting his due credit.
* Will add some new citations and update the doi asap.

Thanks for your input!
