### 2019-03-12 6d5c132
Adding citation to paper, updating bib file
