+++ paper.bib
@@ -1,3 +1,15 @@
+@article{anovitzCharacterizationAnalysisPorosity2015,
+  langid = {english},
+  title = {Characterization and {{Analysis}} of {{Porosity}} and {{Pore Structures}}},
+  volume = {80},
+  doi = {10.2138/rmg.2015.80.04},
+  number = {1},
+  journaltitle = {Reviews in Mineralogy and Geochemistry},
+  date = {2015-01-01},
+  pages = {61-164},
+  author = {Anovitz, Lawrence M. and Cole, David R.},
+}
+
 @article{gostickVersatileEfficientPore2017,
   title = {Versatile and Efficient Pore Network Extraction Method Using Marker-Based Watershed Segmentation},
   volume = {96},
@@ -96,6 +108,7 @@
   title = {Focused Ion Beam Milling: {{A}} Method of Site-Specific Sample Extraction for Microanalysis of {{Earth}} and Planetary Materials},
   volume = {86},
   number = {9},
+  doi = {10.2138/am-2001-8-917},
   journaltitle = {American Mineralogist},
   date = {2001-09},
   pages = {1094-1099},
