### 2019-03-13 fbce7f0
Adding citations
