+++ paper.bib
@@ -1,3 +1,29 @@
+@article{karanPEFCCatalystLayer2017,
+  title = {{{PEFC}} Catalyst Layer: {{Recent}} Advances in Materials, Microstructural Characterization, and Modeling},
+  volume = {5},
+  issn = {2451-9103},
+  url = {http://www.sciencedirect.com/science/article/pii/S2451910317300406},
+  doi = {10.1016/j.coelec.2017.08.018},
+  number = {1},
+  journaltitle = {Current Opinion in Electrochemistry},
+  date = {2017-10-01},
+  pages = {27-35},
+  author = {Karan, Kunal},
+}
+
+@article{kumarReservoirSimulationCO22005,
+  langid = {english},
+  title = {Reservoir {{Simulation}} of {{CO}}{\textsubscript{2}} {{Storage}} in {{Aquifers}}},
+  volume = {10},
+  url = {https://www.onepetro.org/journal-paper/SPE-89343-PA},
+  doi = {10.2118/89343-PA},
+  number = {03},
+  journaltitle = {SPE Journal},
+  date = {2005-09-01},
+  pages = {336-348},
+  author = {Kumar, Ajitabh and Noh, Myeong H. and Ozah, Robin C. and Pope, Gary A. and Bryant, Steven L. and Sepehrnoori, Kamy and Lake, Larry W.},
+}
+
 @article{anovitzCharacterizationAnalysisPorosity2015,
   langid = {english},
   title = {Characterization and {{Analysis}} of {{Porosity}} and {{Pore Structures}}},
