### 2019-03-13 b4d6eb5
Fix deprecation warning regarding tuple indexing
