### 2019-03-13 c4c82ce
Improve overlapping_spheres: porosity of images now accurate down to 1%
