### 2019-03-13 a223964
Improve polydisperse_spheres: image porosity now accurate down to 1-3%
