### 2019-03-13 d607fad
Updated tests for sphere generators to cover a range of porosity
