### 2019-03-13 3f40f45
Changed the default tolerance for overlapping_spheres to 0.02
