### 2019-03-13 bf5aa73
Changed some inside parameters for overlapping_spheres for robustness
