+++ test/unit/test_generators.py
@@ -38,18 +38,18 @@ class GeneratorTest():
     def test_overlapping_spheres_2d(self):
         phis = sp.arange(0.1, 0.9, 0.2)
         for phi in phis:
-            im = ps.generators.overlapping_spheres(shape=[101, 101], radius=6,
-                                                    porosity=phi)
+            im = ps.generators.overlapping_spheres(shape=[101, 101], radius=5,
+                                                   porosity=phi)
             phi_actual = im.sum() / sp.size(im)
-            assert abs(phi_actual - phi) < 0.01
+            assert abs(phi_actual - phi) < 0.02
 
     def test_overlapping_spheres_3d(self):
         phis = sp.arange(0.1, 0.9, 0.2)
         for phi in phis:
             im = ps.generators.overlapping_spheres(shape=[100, 100, 50],
-                                                    radius=8, porosity=phi)
+                                                   radius=8, porosity=phi)
             phi_actual = im.sum() / sp.size(im)
-            assert abs(phi_actual - phi) < 0.01
+            assert abs(phi_actual - phi) < 0.02
 
     def test_polydisperse_spheres(self):
         phis = sp.arange(0.1, 0.9, 0.2)
@@ -59,47 +59,47 @@ class GeneratorTest():
                                                     porosity=phi, dist=dist,
                                                     nbins=10)
             phi_actual = im.sum() / sp.size(im)
-            assert abs(phi_actual - phi) < 0.03
+            assert abs(phi_actual - phi) < 0.05
 
     def test_voronoi_edges(self):
         sp.random.seed(0)
         im = ps.generators.voronoi_edges(shape=[50, 50, 50],
-                                          radius=2,
-                                          ncells=25,
-                                          flat_faces=True)
+                                         radius=2,
+                                         ncells=25,
+                                         flat_faces=True)
         top_slice = im[:, :, 0]
         assert sp.sum(top_slice) == 1409
 
     def test_lattice_spheres_square(self):
         im = ps.generators.lattice_spheres(shape=[101, 101], radius=5,
-                                            offset=0, lattice='sc')
+                                           offset=0, lattice='sc')
         labels, N = spim.label(input=~im)
         assert N == 100
 
     def test_lattice_spheres_triangular(self):
         im = ps.generators.lattice_spheres(shape=[101, 101], radius=5,
-                                            lattice='triangular')
+                                           lattice='triangular')
         labels, N = spim.label(input=~im)
         assert N == 85
 
     def test_lattice_spheres_sc(self):
         im = ps.generators.lattice_spheres(shape=[101, 101, 101],
-                                            radius=4, offset=1,
-                                            lattice='sc')
+                                           radius=4, offset=1,
+                                           lattice='sc')
         labels, N = spim.label(input=~im)
         assert N == 1000
 
     def test_lattice_spheres_fcc(self):
         im = ps.generators.lattice_spheres(shape=[101, 101, 101],
-                                            radius=4, offset=2,
-                                            lattice='fcc')
+                                           radius=4, offset=2,
+                                           lattice='fcc')
         labels, N = spim.label(input=~im)
         assert N == 392
 
     def test_lattice_spheres_bcc(self):
         im = ps.generators.lattice_spheres(shape=[101, 101, 101],
-                                            radius=4, offset=2,
-                                            lattice='bcc')
+                                           radius=4, offset=2,
+                                           lattice='bcc')
         labels, N = spim.label(input=~im)
         assert N == 1024
 
@@ -138,7 +138,7 @@ class GeneratorTest():
     def test_RSA_mask_edge_2d(self):
[…]
