### 2019-03-13 2affb72
Better default parameters for overlapping_spheres
