### 2019-03-13 2532cd0
Better parameters for Newton's method for faster convergence/robustness
