### 2019-03-13 5334f7b
Add bisection search, alternative to Newton method: more robust, slower
