### 2019-03-15 8045f50
adding contributors guide
