### 2019-03-15 7d78258
Improve perturbation parameter in Newton's method in sphere generators
