### 2019-03-15 1b82278
Fixing typos
