### 2019-03-15 PMEAL/porespy#151 @yxqd
Improve documentation: images
Refs openjournals/joss-reviews#1296

The images on the documentation site seem to be hosted at i.imgur.com and sometimes they cannot be displayed:

![image](https://user-images.githubusercontent.com/1796155/54443257-a4b9bf80-4716-11e9-9589-71e207a47b92.png)

Please consider moving the images in the [documenation source code](https://github.com/PMEAL/porespy/tree/master/docs) or somewhere else more reliable
