### 2019-03-15 PMEAL/porespy#153 @yxqd
Improve tests
Refs openjournals/joss-reviews#1296

It is great that tests for many algorithms are included. Some tests could be improved to boost user confidence on the algorithms. For example it looks to me that this test https://github.com/PMEAL/porespy/blob/8045f506364e6278c68cf5ebaaff4ef604bc13a1/test/unit/test_filters.py#L170 only checks the maximum of the result. Could you please add some simple tests that contain more validation? For this particular example, maybe you can add a test that

* uses a fixed input image instead of a random image
* compares the result of the filter to a saved, expected result, and makes sure the difference is within a tolerance.
