### 2019-03-15 PMEAL/porespy#154 @yxqd
improve documentation: references or brief descriptions of some algorithms
Refs openjournals/joss-reviews#1296

Could you please consider adding brief descriptions or references for implementations for algorithms. For example the local thickness algorithm:https://github.com/PMEAL/porespy/blob/8045f506364e6278c68cf5ebaaff4ef604bc13a1/porespy/filters/__funcs__.py#L833
The description on what it does is clear, but the implementation is not explained or referenced. A good example is the documentation for the reconstruction algorithms in tomopy: https://tomopy.readthedocs.io/en/latest/api/tomopy.recon.algorithm.html
