### 2019-03-15 PMEAL/porespy#155 @yxqd
(minor) improve README: axis labels
Refs openjournals/joss-reviews#1296

This is a very minor request. Please consider adding axis labels to the line plots in the README.md.
