### 2019-03-15 #1296 @yxqd
@jgostick Thanks! The importing and testing following the instructions in README went well.

This is a software that has a clear-defined audience and lower the barrier for research concerning porosity analysis. Great job! Overall I am impressed with the quality of the software and documentation, and would recommend it for publication pending some issues below. Most of the issues are minor and are only suggestions, but I would really appreciate the authors to address the issue on improving the tests.

* PMEAL/porespy#153
* PMEAL/porespy#154
* PMEAL/porespy#155
