### 2019-03-15 6ef451e
Moving figures to docs folder and updating refers from imgur
