### 2019-03-15 730a9a1
updating contrib guide
