### 2019-03-15 PMEAL/porespy#151 @jgostick
This is now done, figures have been added to docs/_static, and the docs all link there now.  Only 2 references are left to imgur, and these are from jupyter notebooks in the examples folder, which are rendered by github so work ok.
