### 2019-03-15 PMEAL/porespy#155 @jgostick
done and pushed
