### 2019-03-15 258bcdf
adding labels to figures in readme examples
