### 2019-03-15 #1296 @yxqd
@jgostick BTW, it would be really nice to have the binder examples working: https://github.com/PMEAL/porespy-examples. This is just a suggestion though and not a road-blocker.
