### 2019-03-15 #1296 @jgostick
Binder does not seem to work for us.  We got it going at first, but now it wont' render our files.  I think we may have an example that asks for too much resources?  Or maybe we broke the configuration by accident.
