### 2019-03-15 d7e8a34
adding some extra citations to functions
