### 2019-03-15 PMEAL/porespy#154 @jgostick
Many functions already had documentation.  The functions that do not are our own implementations.  For instance, our local thickness does not work the same as the imageJ version.  Nonetheless, I have added references in several places as best I can.
