### 2019-03-15 PMEAL/porespy#154 @yxqd
I think your implementation of local thickness algorithm is quite ingenious. It is fast and simple, although it has a clear limit of accuracy determined by parameter `sizes`. I have to read your documentation and code carefully to understand it. So it would be very helpful for users that some words are added about your own implemenations, like this one, explaining it is different from ImageJ, its performance, its limitation, etc.
