### 2019-03-16 PMEAL/porespy#154 @jgostick
Ingenious is a bit of a stretch, but I'm glad you approve!  I just realized as some point that it was 'porosimetry' without access limitations.  I think it's probably too much to explain in the doc string, and we're trying to keep the "docs" lean so we don't have to maintain too much.  I think a better approach would be an example in the examples repo which I'm touching up right now.  I can link to it from the docstring.  In fact, that is a a GREAT idea for ALL the functions, to link to a proper notebook with plenty of explanation.
