### 2019-03-16 #1296 @cr458
@jgostick sorry for the delay in getting back.

1) Are you expecting contributions to this package from external users in the future? If so it may be useful to include instructions for developers too: i.e to install test requirements, and how to run the tests. I see that you include instructions to install the package in editable mode, which is nice.

2) Since in your README you recommend:

> PoreSpy depends heavily on the Scipy Stack. The best way to get a fully functional environment is the Anaconda distribution.

Could you include your recommended way of installing it using conda?

3) Also could you please put:
>Be sure to get the Python 3.6+ version.

in bold/capitals/both, I somehow managed to miss it the first time around and have a default version of 3.5 for some reason (rare case but can happen!)

Will be getting round to digging into the code and playing around with it soon, looks good so far!
