### 2019-03-17 2bb31e0
Updating readme, removing refs to examples repo, fixing pep8 error
