### 2019-03-19 bc0a29b
adding trim_disconnected_blobs to filters
