### 2019-03-19 512ef45
Fixing broken tests, deprecation warnings
