### 2019-03-20 318105a
add pyevtk back in to install requires and stop ignoring it in io if not there
