+++ porespy/io/__funcs__.py
@@ -4,10 +4,7 @@ from scipy import ndimage as spim
 import scipy.ndimage as nd
 from pathlib import Path
 from porespy.networks import generate_voxel_image
-try:
-    from pyevtk.hl import imageToVTK
-except:
-    print("warning: pyevtk must be install manually")
+from pyevtk.hl import imageToVTK
 
 
 def dict_to_vtk(data, path='./dictvtk', voxel_size=1, origin=(0, 0, 0)):
