### 2019-03-20 896f4e0
Activate the snow_advanced notebook test and update basic to no longer align_with_openpnm
