### 2019-03-20 fdd3a41
Handle 2d images in to_vtk and add test
