+++ porespy/io/__funcs__.py
@@ -97,6 +97,8 @@ def to_vtk(im, path='./voxvtk', divide=False, downsample=False, voxel_size=1,
     -----
     Outputs a vtk, vtp or vti file that can opened in paraview
     """
+    if len(im.shape) == 2:
+        im = im[:, :, np.newaxis]
     if im.dtype == bool:
         vox = True
     if vox:
