+++ test/unit/test_examples.py
@@ -83,6 +83,6 @@ if __name__ == '__main__':
     self = t
     t.setup_class()
     for item in t.__dir__():
-        if item.startswith('test_network_extraction_snow_advanced'):
+        if item.startswith('test'):
             print('running test: '+item)
             t.__getattribute__(item)()
