### 2019-03-20 PMEAL/porespy#153 @jgostick
Our approach is to create an identical image at the initialization of the test class (self.im), then use it for all tests.  This saves us having to store a 200**3 image in the repo.  We ensure the image is identical each time to setting the seed of the random number generator then calling ``blobs``.

To make sure this approach is solid I have added an assert statement that checks the porosity.

There were also several recently added tests that did not use ``self.im`` so I changed those.
