### 2019-03-20 4d04c6b
touch-ups to a few tests in filters
