### 2019-03-22 #1296 @cr458
@jgostick I've noticed quite a few missing tests. Would you like me to open separate issues for each one, or one big issue for all of them?
