+++ porespy/networks/__snow__.py
@@ -1,6 +1,6 @@
-from porespy.network_extraction import regions_to_network, add_boundary_regions
-from porespy.network_extraction import _net_dict
-from porespy.network_extraction import label_boundary_cells
+from porespy.networks import regions_to_network, add_boundary_regions
+from porespy.networks import _net_dict
+from porespy.networks import label_boundary_cells
 from porespy.tools import pad_faces
 from porespy.filters import snow_partitioning
 from porespy.tools import make_contiguous
