### 2019-03-28 c32e0ea
alias in snow_partitioning
