+++ porespy/tools/__init__.py
@@ -57,6 +57,7 @@ that do NOT return a modified version of the original image.
 '''
 from .__funcs__ import align_image_with_openpnm
 from .__funcs__ import bbox_to_slices
+from .__funcs__ import _create_alias_map
 from .__funcs__ import extend_slice
 from .__funcs__ import extract_subsection
 from .__funcs__ import extract_cylinder
