+++ porespy/networks/__init__.py
@@ -22,6 +22,7 @@ Contains functions for analysing images as pore networks
 .. autofunction:: generate_voxel_image
 .. autofunction:: label_boundary_cells
 .. autofunction:: map_to_regions
+.. autofunction:: add_phase_interconnections
 .. autofunction:: regions_to_network
 .. autofunction:: snow
 .. autofunction:: snow_dual
@@ -34,8 +35,7 @@ from .__funcs__ import add_boundary_regions
 from .__funcs__ import map_to_regions
 from .__funcs__ import generate_voxel_image
 from .__funcs__ import label_boundary_cells
-from .__funcs__ import _create_alias_map
-from .__funcs__ import connect_network_phases
+from .__funcs__ import add_phase_interconnections
 from .__getnet__ import regions_to_network
 from .__utils__ import _net_dict
 from .__snow__ import snow
