+++ porespy/networks/__funcs__.py
@@ -94,17 +94,23 @@ def add_boundary_regions(regions=None, faces=['front', 'back', 'left',
             regions[:, 0, :] = regions[:, 0, :] + regions.max()
             regions[:, -1, :] = regions[:, -1, :] + regions.max()
             regions[:, :, 0] = (~find_boundaries(regions[:, :, 0],
-                                                 mode='outer'))*regions[:, :, 0]
+                                                 mode='outer')) * regions[:, :,
+                                                                  0]
             regions[:, :, -1] = (~find_boundaries(regions[:, :, -1],
-                                                  mode='outer'))*regions[:, :, -1]
+                                                  mode='outer')) * regions[:, :,
+                                                                   -1]
             regions[0, :, :] = (~find_boundaries(regions[0, :, :],
-                                                 mode='outer'))*regions[0, :, :]
+                                                 mode='outer')) * regions[0, :,
+                                                                  :]
             regions[-1, :, :] = (~find_boundaries(regions[-1, :, :],
-                                                  mode='outer'))*regions[-1, :, :]
+                                                  mode='outer')) * regions[-1,
+                                                                   :, :]
             regions[:, 0, :] = (~find_boundaries(regions[:, 0, :],
-                                                 mode='outer'))*regions[:, 0, :]
+                                                 mode='outer')) * regions[:, 0,
+                                                                  :]
             regions[:, -1, :] = (~find_boundaries(regions[:, -1, :],
-                                                  mode='outer'))*regions[:, -1, :]
+                                                  mode='outer')) * regions[:,
+                                                                   -1, :]
             # -----------------------------------------------------------------
             regions = sp.pad(regions, 2, 'edge')
 
@@ -129,13 +135,13 @@ def add_boundary_regions(regions=None, faces=['front', 'back', 'left',
             regions[:, 0] = regions[:, 0] + regions.max()
             regions[:, -1] = regions[:, -1] + regions.max()
             regions[0, :] = (~find_boundaries(regions[0, :],
-                                              mode='outer'))*regions[0, :]
+                                              mode='outer')) * regions[0, :]
             regions[-1, :] = (~find_boundaries(regions[-1, :],
-                                               mode='outer'))*regions[-1, :]
+                                               mode='outer')) * regions[-1, :]
             regions[:, 0] = (~find_boundaries(regions[:, 0],
-                                              mode='outer'))*regions[:, 0]
+                                              mode='outer')) * regions[:, 0]
             regions[:, -1] = (~find_boundaries(regions[:, -1],
-                                               mode='outer'))*regions[:, -1]
+                                               mode='outer')) * regions[:, -1]
             # -----------------------------------------------------------------
             regions = sp.pad(regions, 2, 'edge')
 
@@ -203,8 +209,9 @@ def _generate_voxel_image(network, pore_shape, throat_shape, max_dim=200,
     # Transform points to satisfy origin at (0, 0, 0)
     xyz0 = xyz.min(axis=0) - delta
     xyz += -xyz0
-    res = (xyz.ptp(axis=0).max() + 2*delta) / max_dim
-    shape = np.rint((xyz.max(axis=0) + delta) / res).astype(int) + 2*extra_clearance
+    res = (xyz.ptp(axis=0).max() + 2 * delta) / max_dim
+    shape = np.rint((xyz.max(axis=0) + delta) / res).astype(
+        int) + 2 * extra_clearance
 
     # Transforming from real coords to matrix coords
     xyz = np.rint(xyz / res).astype(int) + extra_clearance
[…]
