### 2019-03-29 9455052
Dealing with deprecation warnings
