### 2019-03-29 462feb4
Adding tests and updating docstrings
