+++ porespy/generators/__imgen__.py
@@ -611,8 +611,8 @@ def blobs(shape: List[int], porosity: float = 0.5, blobiness: int = 1):
 
     porosity : float
         If specified, this will threshold the image to the specified value
-        prior to returning.  If no value is given (the default), then the
-        scalar noise field is returned.
+        prior to returning.  If ``None`` is given, then the scalar noise
+        field is returned.
 
     blobiness : int or list of ints(default = 1)
         Controls the morphology of the blobs.  A higher number results in
