### 2019-03-30 #1296 @jgostick
@cr458 will make a list of missing unit tests, but otherwise I think we addressed all of @yxqd's comments.
