### 2019-03-30 #1296 @cr458
Going through and making  a list of missing lists, documentation and issues I'm encountering when using the data, should be done fairly soon!
