### 2019-03-31 5d1d9a3
Fixing to_openpnm to use openpnm's writer instead of doing it manually
