### 2019-03-31 PMEAL/porespy#159 @cr458
Missing tests
It seems there are a few incomplete/ missing tests, generally for helper functions etc... Most of them look like they would be fairly easy to write tests for. A couple I've noticed are:

In **Generators**:
-  `line_segment()`
- `_remove_edges`
- `_fit_strel_to_im_3d`

In **Filters:**
- `snow_partitioning`

In **Metrics:**
- `representative_elementary_volume`
- `porosity_profile`
- `chord_counts`

In **Networks:**
- ` test_snow_dual_2d` is commented out, is there a reason for this?
-  It would be great to test the content of the returned `network_dict` for expected values and not just NaN values in the `test_snow_dual_3d` and `test_snow_dual_2d` tests.

In **Tools**:
- ` subdivide`
- `extract_cylinder`

I may have missed some. All in all the coverage is good! It would just be great to cover the missing tests.
