### 2019-03-31 PMEAL/porespy#160 @cr458
noise function
The [`noise`](https://github.com/PMEAL/porespy/blob/8045f506364e6278c68cf5ebaaff4ef604bc13a1/porespy/generators/__imgen__.py#L480) function seems to require the installation of a package.

1. Could we change the name of this function to `generate_noise`, or `get_noise` since it has the same name as the `noise` package.

2. Can we include the `noise` package in `requirements.txt`?
