### 2019-03-31 PMEAL/porespy#161 @cr458
blobs docstring
[`blobs`](https://github.com/PMEAL/porespy/blob/8045f506364e6278c68cf5ebaaff4ef604bc13a1/porespy/generators/__imgen__.py#L565) docstring says the default value of `porosity` is `None` and a 'scalar noise field' is returned, but default value seems to be 0.5?
