### 2019-03-31 PMEAL/porespy#162 @cr458
insert_shape docstring
docstring missing at [`insert_shape`](https://github.com/PMEAL/porespy/blob/8045f506364e6278c68cf5ebaaff4ef604bc13a1/porespy/generators/__imgen__.py#L11)
