### 2019-03-31 PMEAL/porespy#163 @cr458
snow and snow_dual files
More of a suggestion than a rigid issue: would it make sense to combine `snow` and `snow_dual` into the same .py file, and create a separate `Network_Dict` class/namedtuple that both functions had access to and could be returned without defining the class within both functions separately?
