### 2019-03-31 PMEAL/porespy#164 @cr458
Dimension checking
In many functions such as[ `distance_transform_lin`](https://github.com/PMEAL/porespy/blob/8045f506364e6278c68cf5ebaaff4ef604bc13a1/porespy/filters/__funcs__.py#L17), dimension check is performed using `arr.ndim`. However if the user has not 'squeezed' the input array, this may return unexpected results:
```

In [14]: test_arr = np.array([[0, 0, 0]])

In [15]: print('ndims: %s, squeezed ndims: %s' %(test_arr.ndim,test_arr.squeeze().ndim))
ndims: 2, squeezed ndims: 1
```
Would be good to find occurrences of this, and perform the dimension check on the squeezed array, and output a warning when `arr.ndims != arr.squeeze().ndim`
