### 2019-03-31 #1296 @cr458
@jgostick have left some issues on the repo. If most of them seem nitpickiy it's because it's a very nice package and I found no major issues :).

- PMEAL/porespy#164
- PMEAL/porespy#163
- PMEAL/porespy#162
- PMEAL/porespy#161
- PMEAL/porespy#160
- PMEAL/porespy#159
