### 2019-03-31 #1296 @jgostick
Loving it, many thanks.  I will address asap, then merge dev into master, upgrade version number to 1.1 and deploy to PyPI.   I would like that version to be the official JOSS version.  How does that work?
