### 2019-04-01 7e5d031
Changing noise to generate_noise
