### 2019-04-01 PMEAL/porespy#160 @jgostick
Name has been changed, but I don't want to add the package to our requirements since it cannot be installed from pip on windows machines as it needs to be compiled.  I can be installed from conda, so that information is included in the docstring.
