### 2019-04-01 1e3730e
Updating docstring in blobs
