### 2019-04-01 PMEAL/porespy#161 @jgostick
Nice catch...fixed
