### 2019-04-01 264b381
Enhanced insert_shape considerably, and added docstring and unit tests
