### 2019-04-01 PMEAL/porespy#162 @jgostick
Done and enhanced the insert shape function significantly
