### 2019-04-01 c2867a3
extra test for insert_shape
