### 2019-04-01 PMEAL/porespy#163 @jgostick
We've actually done part of this in a branch that was merged to dev last week.  The Network_Dict class is now in a utils file, and is used by snow, snow_dual, and snow_n.  We also refactored these three algorithms to all call the same set of functions.  Regarding combining into same file, I prefer separate files to help keep file length manageable.
