### 2019-04-01 a3deddd
Adding squeeze check warning to various functions
