+++ porespy/filters/__funcs__.py
@@ -3,6 +3,7 @@ import scipy as sp
 import numpy as np
 import scipy.ndimage as spim
 import scipy.spatial as sptl
+import warnings
 from scipy.signal import fftconvolve
 from tqdm import tqdm
 from numba import jit
@@ -48,6 +49,10 @@ def distance_transform_lin(im, axis=0, mode='both'):
         A copy of ``im`` with each foreground voxel containing the distance to
         the nearest background along the specified axis.
     """
+    if im.ndim != im.squeeze().ndim:
+        warnings.warn('Input image conains a singleton axis:' + str(im.shape) +
+                      ' Reduce dimensionality with np.squeeze(im) to avoid' +
+                      ' unexpected behavior.')
     if mode in ['backward', 'reverse']:
         im = sp.flip(im, axis)
         im = distance_transform_lin(im=im, axis=axis, mode='forward')
@@ -333,6 +338,10 @@ def find_peaks(dt, r_max=4, footprint=None):
     faster than using a circular or spherical element.
     """
     im = dt > 0
+    if im.ndim != im.squeeze().ndim:
+        warnings.warn('Input image conains a singleton axis:' + str(im.shape) +
+                      ' Reduce dimensionality with np.squeeze(im) to avoid' +
+                      ' unexpected behavior.')
     if footprint is None:
         if im.ndim == 2:
             footprint = disk
@@ -543,6 +552,10 @@ def find_disconnected_voxels(im, conn=None):
         ``im`` using: ``im[holes] = False``
 
     """
+    if im.ndim != im.squeeze().ndim:
+        warnings.warn('Input image conains a singleton axis:' + str(im.shape) +
+                      ' Reduce dimensionality with np.squeeze(im) to avoid' +
+                      ' unexpected behavior.')
     if im.ndim == 2:
         if conn == 4:
             strel = disk(1)
@@ -644,6 +657,10 @@ def trim_nonpercolating_paths(im, inlet_axis=0, outlet_axis=0):
     trim_blind_pores
 
     """
+    if im.ndim != im.squeeze().ndim:
+        warnings.warn('Input image conains a singleton axis:' + str(im.shape) +
+                      ' Reduce dimensionality with np.squeeze(im) to avoid' +
+                      ' unexpected behavior.')
     im = trim_floating_solid(~im)
     labels = spim.label(~im)[0]
     inlet = sp.zeros_like(im, dtype=int)
@@ -879,6 +896,10 @@ def apply_chords(im, spacing=1, axis=0, trim_edges=True, label=False):
     apply_chords_3D
 
     """
+    if im.ndim != im.squeeze().ndim:
+        warnings.warn('Input image conains a singleton axis:' + str(im.shape) +
+                      ' Reduce dimensionality with np.squeeze(im) to avoid' +
+                      ' unexpected behavior.')
     if spacing < 0:
         raise Exception('Spacing cannot be less than 0')
     if spacing == 0:
@@ -938,6 +959,10 @@ def apply_chords_3D(im, spacing=0, trim_edges=True):
     apply_chords
 
     """
+    if im.ndim != im.squeeze().ndim:
+        warnings.warn('Input image conains a singleton axis:' + str(im.shape) +
+                      ' Reduce dimensionality with np.squeeze(im) to avoid' +
+                      ' unexpected behavior.')
     if im.ndim < 3:
         raise Exception('Must be a 3D image to use this function')
     if spacing < 0:
@@ -1092,6 +1117,10 @@ def porosimetry(im, sizes=25, inlets=None, access_limited=True,
     local_thickness
 
     """
+    if im.ndim != im.squeeze().ndim:
+        warnings.warn('Input image conains a singleton axis:' + str(im.shape) +
[…]
