### 2019-04-01 PMEAL/porespy#164 @jgostick
Good point, warning has been added throughout
