### 2019-04-01 132e7f8
adding tests for line_segments, improved to 2D and 3D
