### 2019-04-01 b722ad7
Adding test for chord_counts
