### 2019-04-01 d36a064
Adding test for extract_cylinder, updating code.
