### 2019-04-01 4b78018
Add tests for subdivide
