### 2019-04-01 PMEAL/porespy#159 @jgostick
Done, done and done!
