### 2019-04-02 cc79189
Fixing docstring for cylinders, and improving functionality
