### 2019-04-02 #1296 @usethedata
@jgostick I've double checked, and the answer to your question is that when we're ready, you make a deposit of the desired version (such as with Zenodo).  Put the DOI for that archive here in this thread, and I'll use that to set the archive version associated with the paper.
