### 2019-04-06 d1c8cc1
scikit-image 0.15 added exceptions, so we now handle them
