### 2019-04-06 7f5adcd
Fixing a few tests, and updating keywords in examples
