### 2019-04-08 459bdd0
Updating version number to 1.1.0
