+++ porespy/__init__.py
@@ -79,7 +79,7 @@ Use some filters from PoreSpy:
 
 '''
 
-__version__ = "1.0.1"
+__version__ = "1.1.0"
 
 from . import tools
 from . import filters
