### 2019-04-08 #1296 @jgostick
Hello all:  I have now merged everything onto master, bumped the version number to V1.1.0, and pushed official release to PyPI.  I have also enabled Zenodo, and received the following DOI for this release:  10.5281/zenodo.2633285
