### 2019-04-08 6176201
Switched default solver in sphere generator to bisection method
