### 2019-04-10 #1296 @jgostick
I have fixed my release debacle, and now version 1.1.1 of PoreSpy is available on Github.  It now has a DOI of 10.5281/zenodo.2635591 for V1.1.1.
