### 2019-04-10 83183b2
Added zenodo badge and bumping version number
