### 2019-04-12 #1296 @jgostick
Sorry about that @TomTranter.  There is also another issue:  The Tag on github is 1.1.1, and this has been archived with a DOI...but the __version__ number is still at 1.1.0  👎   I really need to learn how to do githooks to prevent this, because I do it too often.  So how to proceed?  Should I add a new tag and have a new doi issued?
