### 2019-04-14 d47deea
Updating readme to include DOI badge, incrementing version to 1.1.2
