### 2019-04-14 fb0b26b
Updating readme to include doi and citation
