### 2019-04-14 ec4de18
format citation better
