+++ README.rst
@@ -15,17 +15,13 @@
    :target: https://www.doi.org/10.5281/zenodo.2633284
 
 
-::
 
-    Cite as:
-
-    Gostick J, Khan ZA, Tranter TG, Kok MDR, Agnaou M, Sadeghi MA, Jervis R.
-
-    PoreSpy: A Python Toolkit for Quantitative Analysis of Porous Media Images.
-
-    Journal of Open Source Software, 2019.
-
-    doi:10.5281/zenodo.2633284
+|  Cite as:
+|
+|  Gostick J, Khan ZA, Tranter TG, Kok MDR, Agnaou M, Sadeghi MA, Jervis R.
+|  PoreSpy: A Python Toolkit for Quantitative Analysis of Porous Media Images.
+|  Journal of Open Source Software, 2019.
+|  doi:10.5281/zenodo.2633284
 
 
 -------------------------------------------------------------------------------
