### 2019-04-15 #1296 @jgostick
Great!  I have not 'pulled the trigger' on this new doi/tag/version yet due to mistakes made previously.  The dev branch is all set to go.  Now that both reviewers have reported back positively, I suppose I can:

1. Merge dev onto master
2. Tag master as V1.1.2 (Will be the official version of record in JOSS, also I can assure you that __version__ will now be set to 1.1.2 as well!)
3.  This tagging will generate a new DOI, which I will paste back into here when I get it.

Is this correct?
