### 2019-04-16 #1296 @usethedata
@cr458 If you'll mark off the rest of your checklist, that would be good.
@jgostick Yes.  I'll change the version here and I'll update the DOI when you've got that.
