### 2019-04-16 #1296 @jgostick
This DOI is offered on Zenodo as the 'one to rule them all', which will always resolve to the latest release.  I feel like this is the best one to use as future versions won't render it obsolete...

10.5281/zenodo.2633284
