### 2019-04-28 cb54a67
Add additional project metadata to setup function
