+++ setup.py
@@ -52,5 +52,10 @@ setup(
                       'openpnm'],
     author='Jeff Gostick',
     author_email='jgostick@gmail.com',
-    url='http://porespy.org'
+    url='http://porespy.org',
+    project_urls={
+        'Documentation': 'https://porespy.readthedocs.io/en/master/',
+        'Source': 'https://porespy.readthedocs.io/en/master/',
+        'Tracker': 'https://github.com/PMEAL/porespy/issues',
+    },
 )
