### 2019-04-30 #1296 @jgostick
Hi All, what's the situation here?  I think everything is done on our end.  I would like be able to 'announce' this paper at the upcoming interpore conference in Valencai, which is May 6-10...assuming our submission is acceptable, will it be possible to make a decision by this date?
