### 2019-04-30 #1296 @usethedata
@jgostick My apologies for the delay.  Some elder care issues have needed my attention.

@openjournals/joss-eics I believe this is ready to go.
