### 2019-04-30 #1296 @usethedata
@jgostick If you want to give the PDF one final review to make sure you don't see any mistakes (author list, etc), that would be good.  It's a pain to fix once the article has been published.
