### 2019-05-01 #1296 @kyleniemeyer
Hi @jgostick, before I do the final accept, could you edit the metadata for the Zenodo archive so that the authors match the paper? And also add full names/ORCID rather than just GitHub handles.
