### 2019-05-01 #1296 @jgostick
I think it's done now...hopefully I edited the right places
