### 2019-05-01 #1296 @kyleniemeyer
@jgostick this is what I'm seeing:

In addition to the GitHub handles instead of names, I also see names on Zenodo that aren't on the paper, and vice versa.

On the other hand, for version 1.1.1 in Zenodo, I see this:

So perhaps you edited the names for the wrong version?
