### 2019-05-01 #1296 @jgostick
ahhh, I clicked save, but not publish, because I thought it already was published.

Names, affiliation and orcid should be ok now, but I'm a bit confused which DOI to use...

The DOI for V1.1.2 is 10.5281/zenodo.2642818, but zenodo offers one for "all versions" which seems more useful since a new doi is issued for every minor version bump.  The one for all versions is 10.5281/zenodo.2633284, which is the one I gave to @usethedata in the above thread.
