### 2019-05-01 #1296 @kyleniemeyer
In this case, we should use the specific DOI associated with the version published—we don't want the archived software referenced in the paper to point to a future version, if you update the archive later.

The paper also points to the software repo, so it will direct to the latest version in that way.
