### 2019-05-01 #1296 @kyleniemeyer
Congrats @jgostick on your paper's acceptance in JOSS! Thanks to @yxqd and @cr458 for reviewing, and @usethedata for editing.
